backend/requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
sqlmodel>=0.0.16
google-generativeai>=0.5.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
json-logic-qubit>=0.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.14.0
